requirements.txt
fastapi
uvicorn
pyjwt
sqlmodel
werkzeug
cryptography
pydantic-settings